+++ TestUtilities/Mocks/CoreMocks/PassthroughCoreMock.swift
@@ -86,10 +86,9 @@ open class PassthroughCoreMock: DatadogCoreProtocol, FeatureScope {
     public func get<T>(feature type: T.Type) -> T? where T: DatadogFeature { nil }
 
     /// Always returns a feature-scope.
-    public func scope(for feature: String) -> FeatureScope? {
+    public func scope<T>(for featureType: T.Type) -> FeatureScope where T : DatadogFeature {
         self
     }
-
 
     public func set(baggage: @escaping () -> FeatureBaggage?, forKey key: String) {
         context.baggages[key] = baggage()
@@ -116,6 +115,8 @@ open class PassthroughCoreMock: DatadogCoreProtocol, FeatureScope {
     public func context(_ block: @escaping (DatadogContext) -> Void) {
         block(context)
     }
+
+    public var dataStore: DataStore { NOPDataStore() }
 
     /// Recorded events from feature scopes.
     ///